oracle/requirements.txt
web3
flake8
py-solc-x
pytest
oasis-sapphire-py
httpx==0.27.0
python-dotenv==1.0.1
pycryptodome==3.20.0
